Move related UI class to a new place
